requirements.txt
flask
flask-sqlalchemy
Mastodon.py
apscheduler
psycopg2
